refactor: centralize messaging, sharing and voice
