+++ lib/main.dart
@@ -17,7 +17,7 @@ import 'package:anisphere/modules/noyau/providers/user_provider.dart';
 import 'package:anisphere/modules/noyau/providers/animal_provider.dart';
 import 'package:anisphere/modules/noyau/providers/ia_context_provider.dart';
 import 'package:anisphere/modules/noyau/providers/support_provider.dart';
-import 'package:anisphere/modules/messagerie/providers/messaging_provider.dart';
+import 'package:anisphere/modules/noyau/providers/messaging_provider.dart';
 import 'package:anisphere/modules/noyau/services/notification_service.dart';
 import 'package:anisphere/modules/noyau/services/offline_sync_queue.dart';
 import 'package:anisphere/modules/noyau/logic/ia_metrics_collector.dart';
@@ -25,9 +25,9 @@ import 'package:anisphere/modules/noyau/services/cloud_notification_listener.dar
 import 'package:anisphere/modules/noyau/services/navigation_service.dart';
 import 'package:anisphere/modules/noyau/logic/ia_master.dart';
 import 'package:anisphere/modules/noyau/models/support_ticket_model.dart';
-import 'package:anisphere/modules/messagerie/models/message_model.dart';
-import 'package:anisphere/modules/messagerie/models/conversation_model.dart';
-import 'package:anisphere/modules/messagerie/services/offline_message_queue.dart';
+import 'package:anisphere/modules/noyau/models/message_model.dart';
+import 'package:anisphere/modules/noyau/models/conversation_model.dart';
+import 'package:anisphere/modules/noyau/services/offline_message_queue.dart';
 import 'package:anisphere/modules/noyau/providers/photo_provider.dart';
 import 'package:anisphere/modules/noyau/models/photo_model.dart';
 import 'package:anisphere/modules/noyau/models/sync_metrics_model.dart';
