--- lib/modules/messagerie/README.md
@@ -1,3 +1,0 @@
-# Module messagerie
-
-Ce dossier contient le code source du module `messagerie`.
