+++ lib/modules/noyau/logic/ia_message_analyzer.dart
