+++ lib/modules/noyau/models/conversation_model.dart
