+++ lib/modules/noyau/services/messaging_service.dart
