+++ lib/modules/noyau/services/offline_message_queue.dart
