--- lib/modules/sharing/README.md
@@ -1,3 +1,0 @@
-# Module sharing
-
-Ce dossier contient le code source du module `sharing`.
