--- lib/modules/sharing/services/local_sharing_service.dart
@@ -1,14 +1,0 @@
-library;
-
-/// Service local de partage pour AniSphère.
-class LocalSharingService {
-  final Set<String> _sharedIds = {};
-
-  /// Ajoute un identifiant partagé localement.
-  void share(String animalId) {
-    _sharedIds.add(animalId);
-  }
-
-  /// Vérifie si l'identifiant a déjà été partagé.
-  bool isShared(String animalId) => _sharedIds.contains(animalId);
-}
