--- lib/modules/sharing/services/sharing_ia_optimizer.dart
@@ -1,9 +1,0 @@
-library;
-
-/// Optimiseur IA pour les opérations de partage.
-class SharingIAOptimizer {
-  /// Retourne une version optimisée de la chaîne fournie.
-  static String optimize(String input) {
-    return input.trim().toLowerCase();
-  }
-}
