+++ test/noyau/unit/offline_message_queue_test.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
-import 'package:anisphere/modules/messagerie/models/message_model.dart';
-import 'package:anisphere/modules/messagerie/services/offline_message_queue.dart';
+import 'package:anisphere/modules/noyau/models/message_model.dart';
+import 'package:anisphere/modules/noyau/services/offline_message_queue.dart';
 import '../../test_config.dart';
 
 void main() {
