+++ test/noyau/widget/message_list_screen_widget_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
-import 'package:anisphere/modules/messagerie/screens/message_list_screen.dart';
-import 'package:anisphere/modules/messagerie/models/conversation_model.dart';
-import 'package:anisphere/modules/messagerie/providers/messaging_provider.dart';
-import 'package:anisphere/modules/messagerie/services/messaging_service.dart';
+import 'package:anisphere/modules/noyau/screens/message_list_screen.dart';
+import 'package:anisphere/modules/noyau/models/conversation_model.dart';
+import 'package:anisphere/modules/noyau/providers/messaging_provider.dart';
+import 'package:anisphere/modules/noyau/services/messaging_service.dart';
 import '../../test_config.dart';
 
 class _FakeService extends MessagingService {}
