+++ test/noyau/widget/voice_input_button_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:anisphere/modules/voice/voice_input_button.dart';
+import 'package:anisphere/modules/noyau/widgets/voice_input_button.dart';
 
 void main() {
   testWidgets('button shows mic icon', (tester) async {
