+++ test/test_tracker.md
@@ -87,17 +87,17 @@
 | test/noyau/widget/settings_screen_test.dart | widget | package:anisphere/modules/noyau/screens/settings_screen.dart | ✅ |
 | test/noyau/widget/legal_screen_test.dart | widget | package:anisphere/modules/noyau/screens/legal_screen.dart | ✅ |
 | test/noyau/unit/sync_metrics_model_test.dart | unit | package:anisphere/modules/noyau/models/sync_metrics_model.dart | ✅ |
-| test/messagerie/unit/messaging_service_test.dart | unit | package:anisphere/modules/messagerie/services/messaging_service.dart | ✅ |
-| test/messagerie/unit/offline_message_queue_test.dart | unit | package:anisphere/modules/messagerie/services/offline_message_queue.dart | ✅ |
-| test/messagerie/unit/ia_message_analyzer_test.dart | unit | package:anisphere/modules/messagerie/services/ia_message_analyzer.dart | ✅ |
-| test/messagerie/widget/chat_screen_widget_test.dart | widget | package:anisphere/modules/messagerie/screens/chat_screen.dart | ✅ |
-| test/messagerie/widget/message_list_screen_widget_test.dart | widget | package:anisphere/modules/messagerie/screens/message_list_screen.dart | ✅ |
-| test/messagerie/integration/messagerie_integration.dart | integration | package:anisphere/modules/messagerie/screens/chat_screen.dart | ✅ |
+| test/noyau/unit/messaging_service_test.dart | unit | package:anisphere/modules/noyau/services/messaging_service.dart | ✅ |
+| test/noyau/unit/offline_message_queue_test.dart | unit | package:anisphere/modules/noyau/services/offline_message_queue.dart | ✅ |
+| test/noyau/unit/ia_message_analyzer_test.dart | unit | package:anisphere/modules/noyau/logic/ia_message_analyzer.dart | ✅ |
+| test/noyau/widget/chat_screen_widget_test.dart | widget | package:anisphere/modules/noyau/screens/chat_screen.dart | ✅ |
+| test/noyau/widget/message_list_screen_widget_test.dart | widget | package:anisphere/modules/noyau/screens/message_list_screen.dart | ✅ |
+| test/noyau/integration/messagerie_integration.dart | integration | package:anisphere/modules/noyau/screens/chat_screen.dart | ✅ |
 | test/noyau/widget/consent_hooks_test.dart | widget | package:anisphere/modules/noyau/hooks/consent_hooks.dart | ✅ |
-| test/sharing/unit/local_sharing_service_test.dart | unit | package:anisphere/modules/sharing/services/local_sharing_service.dart | ✅ |
-| test/sharing/unit/cloud_sharing_service_test.dart | unit | package:anisphere/modules/sharing/services/cloud_sharing_service.dart | ✅ |
-| test/sharing/unit/sharing_ia_optimizer_test.dart | unit | package:anisphere/modules/sharing/services/sharing_ia_optimizer.dart | ✅ |
-| test/sharing/widget/share_screen_test.dart | widget | package:anisphere/modules/noyau/screens/share_screen.dart | ✅ |
-| test/voice/unit/speech_recognition_service_test.dart | unit | package:anisphere/modules/voice/speech_recognition_service.dart | ✅ |
-| test/voice/unit/voice_command_analyzer_test.dart | unit | package:anisphere/modules/voice/voice_command_analyzer.dart | ✅ |
-| test/voice/widget/voice_input_button_test.dart | widget | package:anisphere/modules/voice/voice_input_button.dart | ✅ |
+| test/noyau/unit/local_sharing_service_test.dart | unit | package:anisphere/modules/noyau/services/local_sharing_service.dart | ✅ |
+| test/noyau/unit/cloud_sharing_service_test.dart | unit | package:anisphere/modules/noyau/services/cloud_sharing_service.dart | ✅ |
+| test/noyau/unit/sharing_ia_optimizer_test.dart | unit | package:anisphere/modules/noyau/services/sharing_ia_optimizer.dart | ✅ |
+| test/noyau/widget/share_screen_test.dart | widget | package:anisphere/modules/noyau/screens/share_screen.dart | ✅ |
+| test/noyau/unit/speech_recognition_service_test.dart | unit | package:anisphere/modules/noyau/services/speech_recognition_service.dart | ✅ |
+| test/noyau/unit/voice_command_analyzer_test.dart | unit | package:anisphere/modules/noyau/logic/voice_command_analyzer.dart | ✅ |
+| test/noyau/widget/voice_input_button_test.dart | widget | package:anisphere/modules/noyau/widgets/voice_input_button.dart | ✅ |
